+++ cxx_mkspecs/gxx_mkspecs.py
@@ -3,15 +3,7 @@
 
 from waflib.Configure import conf
 
-import gxx_common
-
-
-@conf
-def cxx_android_gxx46_arm(conf):
-    """
-    Detect and setup the Android g++ 4.6 compiler for ARM
-    """
-    conf.mkspec_gxx_android_configure(4, 6, 'arm-linux-androideabi')
+from . import gxx_common
 
 
 @conf
@@ -42,68 +34,6 @@ def cxx_android_gxx48_armv7(conf):
 
 
 @conf
-def cxx_crosslinux_gxx46_arm(conf):
-    """
-    Detect and setup the g++ 4.6 cross-compiler for ARM 32-bit Linux
-    """
-    conf.mkspec_gxx_configure(4, 6, 'arm-openwrt-linux')
-    # Note: libstdc++ might not be available on the target platform
-    # Statically link the GCC runtime and the C++ standard library
-    conf.env['LINKFLAGS'] += ['-static-libgcc', '-static-libstdc++']
-    # The GCC runtime does not contain the C++ exception handling functions,
-    # so libgcc_eh.a should also be statically linked
-    conf.env['STLIB'] += ['gcc_eh']
-
-
-@conf
-def cxx_crosslinux_gxx47_arm(conf):
-    """
-    Detect and setup the g++ 4.7 cross-compiler for ARM 32-bit Linux
-    """
-    conf.mkspec_gxx_configure(4, 7, 'arm-openwrt-linux')
-    # Note: libstdc++ might not be available on the target platform
-    # Statically link the GCC runtime and the C++ standard library
-    conf.env['LINKFLAGS'] += ['-static-libgcc', '-static-libstdc++']
-    # The GCC runtime does not contain the C++ exception handling functions,
-    # so libgcc_eh.a should also be statically linked
-    conf.env['STLIB'] += ['gcc_eh']
-
-
-@conf
-def cxx_crosslinux_gxx46_x64(conf):
-    """
-    Detect and setup the g++ 4.6 cross-compiler for 64-bit Linux
-    """
-    conf.mkspec_gxx_configure(4, 6, 'crosslinux-gxx46-x64')
-    conf.mkspec_add_common_flag('-m64')
-    # Note: libstdc++ might not be available on the target platform
-    # Statically link the GCC runtime and the C++ standard library
-    conf.env['LINKFLAGS'] += ['-static-libgcc', '-static-libstdc++']
-
-
-@conf
-def cxx_crosslinux_gxx46_x86(conf):
-    """
-    Detect and setup the g++ 4.6 cross-compiler for 32-bit Linux
-    """
-    conf.mkspec_gxx_configure(4, 6, 'crosslinux-gxx46-x86')
-    conf.mkspec_add_common_flag('-m32')
-    # Note: libstdc++ might not be available on the target platform
-    # Statically link the GCC runtime and the C++ standard library
-    conf.env['LINKFLAGS'] += ['-static-libgcc', '-static-libstdc++']
-
-
-@conf
-def cxx_crosslinux_gxx47_mips(conf):
-    """
-    Detect and setup the g++ 4.7 cross-compiler for MIPS 32-bit Linux
-    """
-    conf.mkspec_gxx_configure(4, 7, 'crosslinux-gxx47-mips')
-    # Statically link in the C++ standard library
-    conf.env['LINKFLAGS'] += ['-static-libstdc++']
-
-
-@conf
 def cxx_gcov_gxx49_x64(conf):
     """
     Configure g++ 4.9 (64 bit) for coverage analysis with gcov
@@ -122,42 +52,6 @@ def cxx_gcov_gxx49_x64(conf):
              '-fno-default-inline']
     conf.env['CFLAGS'] += flags
     conf.env['CXXFLAGS'] += flags
-
-
-@conf
-def cxx_gxx46_x64(conf):
-    """
-    Detect and setup the g++ 4.6 compiler for 64 bit
-    """
-    conf.mkspec_gxx_configure(4, 6)
-    conf.mkspec_add_common_flag('-m64')
-
-
-@conf
-def cxx_gxx46_x86(conf):
-    """
-    Detect and setup the g++ 4.6 compiler for 32 bit linux
-    """
-    conf.mkspec_gxx_configure(4, 6)
-    conf.mkspec_add_common_flag('-m32')
-
-
-@conf
-def cxx_gxx47_x64(conf):
-    """
-    Detect and setup the g++ 4.7 compiler for 64 bit
-    """
-    conf.mkspec_gxx_configure(4, 7)
-    conf.mkspec_add_common_flag('-m64')
-
-
-@conf
-def cxx_gxx47_x86(conf):
-    """
-    Detect and setup the g++ 4.7 compiler for 32 bit
-    """
-    conf.mkspec_gxx_configure(4, 7)
-    conf.mkspec_add_common_flag('-m32')
 
 
 @conf
@@ -212,14 +106,6 @@ def cxx_gxx52_x86(conf):
     """
     conf.mkspec_gxx_configure(5, 2)
     conf.mkspec_add_common_flag('-m32')
-
-
-@conf
-def cxx_raspberry_gxx47_arm(conf):
-    """
-    Detect and setup the g++ 4.7 cross-compiler for Raspberry Pi (Linux)
-    """
-    conf.mkspec_gxx_configure(4, 7, 'raspberry-gxx47-arm')
 
 
 @conf